requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose==3.3.0
cryptography==41.0.1
argon2-cffi==21.3.0